Вынес создание задач и фильтров в отдельные модули
